Wrapped proxy's modify response function in a mutex
We use the proxy's modify response functionality to modify responses on
their return trip to clients. This allows us to modify status codes and
body data as configured by the client.

However, this function is a field on the proxy and needs to be able to
operate on all concurrent requests. Much of the matching logic for the
response modifiers depends on matching on the request body, which is not
possible when handling a response.

As a result, we need to make sure we only modify a single response at a
time, even when the proxy is subject to concurrent load. The simplest
way of ensuring this is to wrap changes to the response modifier
function in a mutex, which fixes that test case also included in this
commit (which highlights the concurrency problems).