feat: Add interaction not found error
The interactiond details endpoint previously replied with a 400 if an interaction was not found.
This is changed to a 404.

PactProxy now returns a defined error when a 404 is received from that endpoint, so that dependent tests
can verify that the details were not returned due to the interaction not existing.

This allows tests using pact proxy to verify that a certain pact interaction was not registered with pact-proxy
and gives confidence that the endpoint was not invoked.