fix: concurrency issue with attempts counter
For each interaction, we keep track of the number of times it has been
called.
That counter can be used in the modifiers to change the response
returned to the client. For example it is possible to change the return
status on the 2nd invocation of a specific interaction.

The code has a bug when concurrent requests happen for the same
interaction.
The change modifies the way the concurrency is managed just for the
request counter, using channels instead of Lock.

As part of this change, we have added a new test that can be used to
prove that the original code has the issue and the proposed version
fixes it.

Co-authored-by: Shreya Garge <shreya.garge@form3.tech>