fix: enable adding multiple modifiers to the same interaction
There was a bug which meant that modifiers based on attempt count would
be overwitten because they were stored with the same key in the
modifiers map. We fix this by adding the attempt count to the key so
that now the keys are different and multiple modifiers can be added for
the same interaction based on attempt.

Co-authored-by: Andrea Rosa <andrea.rosa@form3.tech>