fix empty SERVER_ADDRESS config loading
After bumping packages, envconfig behaviour has changed. If a struct contains a pointer, and no value is being loaded, it will now generate the zero value for the pointed type rather than nil.

This prevented the proxy from being started if no SERVER_ADDRESS was specified, as ConfigureProxy checks for no server address being specified by comparing it to url.URL{}, which because of the pointer is no longer equal.

The fix is a simple config change for envconfig so that it will use nil rather than the zero value for the pointer when no value is specified.