fix: fix loading values from source last req
This will fix a problem with fetching the values from the source
interaction last request. The problem occurs when the values from the
source interaction are being used as a constraint.
Due to the bug the json path library is unable to fetch the specified
values from the source last request document.